Refactor crashhandler
