+++ engine/common/host.c
@@ -846,21 +846,7 @@ void Host_InitCommon( int argc, const char** argv, const char *progname, qboolea
 
 	if( !Sys_CheckParm( "-noch" ) )
 	{
-#ifdef _WIN32
-		SetErrorMode( SEM_FAILCRITICALERRORS );	// no abort/retry/fail errors
-		host.oldFilter = SetUnhandledExceptionFilter( Sys_Crash );
-		host.hInst = GetModuleHandle( NULL );
-#elif !defined (CRASHHANDLER)
-//TODO
-#else
-		struct sigaction act;
-		act.sa_sigaction = Sys_Crash;
-		act.sa_flags = SA_SIGINFO | SA_ONSTACK;
-		sigaction(SIGSEGV, &act, &host.oldFilter);
-		sigaction(SIGABRT, &act, &host.oldFilter);
-		sigaction(SIGBUS, &act, &host.oldFilter);
-		sigaction(SIGILL, &act, &host.oldFilter);
-#endif
+		Sys_SetupCrashHandler();
 	}
 
 	host.change_game = bChangeGame;
